+++ src/components/NewTimeEntryInput.vue
@@ -5,8 +5,8 @@
   <input
     ref="input"
     type="text"
-    projectName="newTimeEntry"
-    class="text-input"
+    project-name="newTimeEntry"
+    class="w-full h-full py-3 text-gray-900 placeholder-gray-400 bg-transparent rounded-md focus:outline-none focus:ring-0"
     :placeholder="t('actionBar.what_are_you_working_on')"
     v-bind="$attrs"
     @input="handleInput"
@@ -14,142 +14,207 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, Ref, ref, onMounted, onBeforeUnmount } from "vue";
+import {
+  defineComponent,
+  Ref,
+  ref,
+  onMounted,
+  onBeforeUnmount,
+  computed,
+  PropType,
+} from "vue";
 import { useI18n } from "vue-i18n";
 import autocomplete, { AutocompleteResult } from "autocompleter";
-import { TimeEntry } from "@/api/types";
+import { Customer, Project, TimeEntry } from "@/api/types";
 import TimeEntriesService from "@/api/TimeEntriesService";
-import { createPopper, Instance as PopperInstance } from "@popperjs/core";
-
-interface TimeEntryAutocompleteItem {
-  label: string;
-  value: TimeEntry;
-}
+import ProjectsService from "@/api/ProjectsService";
+import CustomersService from "@/api/CustomersService";
+import { getCustomerName } from "@/utils/getCustomerName";
+import { TaskData } from "@/api/types";
+import {
+  AutocompleteItem,
+  CustomerAutocompleteItem,
+  ProjectAutocompleteItem,
+  renderAutocompleteCustomerEntry,
+  renderAutocompleteProjectEntry,
+  renderAutocompleteTimeEntryEntry,
+  TimeEntryAutocompleteItem,
+} from "@/utils/renderAutocompleteItem";
 
 export default defineComponent({
   projectName: "NewTimeEntryInput",
   props: {
-    modelValue: { type: String, required: true },
+    modelValue: {
+      type: Object as PropType<TaskData>,
+      required: true,
+    },
   },
-  emits: ["update:modelValue", "time-entry-selected"],
-  setup(_, { emit }) {
+  emits: [
+    "update:modelValue",
+    "time-entry-selected",
+    "customer-selected",
+    "project-selected",
+  ],
+  setup(props, { emit }) {
     const { t } = useI18n();
     const input = ref() as Ref<HTMLInputElement>;
+    let spliceIndex: number | undefined;
+    let autocompleteInstance: AutocompleteResult;
+
+    const modelProxy = computed({
+      get() {
+        return props.modelValue;
+      },
+      set(value: TaskData) {
+        emit("update:modelValue", value);
+      },
+    });
 
     function handleInput(event: Event): void {
       const target = event.target as HTMLInputElement;
-      emit("update:modelValue", target.value);
+      modelProxy.value.description = target.value;
     }
 
-    let autocompleteInstance: AutocompleteResult;
-    let popper: PopperInstance;
+    async function suggestTimeEntries(query: string) {
+      const {
+        data: { data: timeEntries },
+      } =
+        query === ""
+          ? await new TimeEntriesService().include("project").suggest()
+          : await new TimeEntriesService().include("project").search(query);
+
+      const newData: TimeEntryAutocompleteItem[] = timeEntries.map(
+        (timeEntry: TimeEntry) => ({
+          type: "time-entry",
+          label: timeEntry.description,
+          value: timeEntry,
+        })
+      );
+
+      return newData;
+    }
+
+    function getQuerySuggestion(text: string, stopWord: string) {
+      return text.match(new RegExp(`${stopWord}(\\w+)`));
+    }
+
+    async function suggestProject(text: string) {
+      const match = getQuerySuggestion(text, "#");
+
+      const query = match ? match[1] ?? "" : "";
+      spliceIndex = match ? match["index"] ?? spliceIndex : spliceIndex;
+
+      const {
+        data: { data: projects },
+      } =
+        query === ""
+          ? await new ProjectsService().suggest()
+          : await new ProjectsService().search(query);
+
+      const newData: ProjectAutocompleteItem[] = projects.map(
+        (project: Project) => ({
+          type: "project",
+          label: project.name,
+          value: project,
+        })
+      );
+
+      return newData;
+    }
+
+    async function suggestCustomer(text: string) {
+      const match = getQuerySuggestion(text, "@");
+
+      const query = match ? match[1] ?? "" : "";
+      spliceIndex = match ? match["index"] ?? spliceIndex : spliceIndex;
+
+      const {
+        data: { data: customers },
+      } =
+        query === ""
+          ? await new CustomersService().suggest()
+          : await new CustomersService().search(query);
+
+      const newData: CustomerAutocompleteItem[] = customers.map(
+        (customer: Customer) => ({
+          type: "customer",
+          label: getCustomerName(customer),
+          value: customer,
+        })
+      );
+
+      return newData;
+    }
 
     onMounted(() => {
-      autocompleteInstance = autocomplete<TimeEntryAutocompleteItem>({
+      autocompleteInstance = autocomplete<AutocompleteItem>({
         input: input.value,
-        emptyMsg: t("timeEntrySelect.noTimeEntriesFound"),
+        emptyMsg: t("actionBar.noResult"),
         minLength: 0,
         debounceWaitMs: 100,
         showOnFocus: true,
         fetch: async (text, callback) => {
-          const {
-            data: { data: customers },
-          } = await new TimeEntriesService().include("project").suggest(text);
-
-          const newData: TimeEntryAutocompleteItem[] = customers.map(
-            (timeEntry: TimeEntry) => ({
-              label: timeEntry.description,
-              value: timeEntry,
-            })
-          );
-
-          callback(newData);
+          if (!text.includes("#") && !text.includes("@")) {
+            return callback(await suggestTimeEntries(text));
+          } else if (text.search("#") > text.search("@")) {
+            return callback(await suggestProject(text));
+          } else {
+            return callback(await suggestCustomer(text));
+          }
         },
-        render(item: TimeEntryAutocompleteItem): HTMLDivElement | undefined {
-          const wrapper = document.createElement("div");
-          const svg = document.createElementNS(
-            "http://www.w3.org/2000/svg",
-            "svg"
-          );
-          const circle = document.createElementNS(
-            "http://www.w3.org/2000/svg",
-            "circle"
-          );
-          const timeEntrySpan = document.createElement("span");
-          const projectWrapper = document.createElement("div");
-          const projectSpan = document.createElement("span");
-
-          const projectName =
-            item.value?.project?.data.name || t("projectIndicator.noProject");
-          const color = item.value?.project?.data.color_hex || "#676767";
-
-          const timeEntryName = item.value.description;
-
-          [
-            "flex",
-            "justify-between",
-            "select-none",
-            "text-gray-900",
-            "text-sm",
-          ].forEach((eleClass) => wrapper.classList.add(eleClass));
-
-          [
-            "flex",
-            "items-center",
-            "text-xs",
-            "select-none",
-          ].forEach((eleClass) => projectWrapper.classList.add(eleClass));
-          projectWrapper.style.color = color;
-
-          svg.setAttribute("width", "10");
-          svg.setAttribute("height", "10");
-          svg.setAttribute("viewBox", "0 0 10 10");
-          svg.setAttribute("fill", "none");
-          svg.setAttribute("xmlns", "http://www.w3.org/2000/svg");
-          svg.classList.add("mr-2");
-
-          circle.setAttribute("cx", "5");
-          circle.setAttribute("cy", "5");
-          circle.setAttribute("r", "5");
-          circle.classList.add("fill-current");
-
-          timeEntrySpan.textContent = timeEntryName;
-          timeEntrySpan.classList.add("block");
-
-          projectSpan.classList.add("block");
-          projectSpan.textContent = projectName;
-
-          svg.appendChild(circle);
-          projectWrapper.appendChild(svg);
-          projectWrapper.appendChild(projectSpan);
-
-          wrapper.appendChild(timeEntrySpan);
-          wrapper.appendChild(projectWrapper);
-
-          return wrapper;
-        },
-        onSelect(item: TimeEntryAutocompleteItem) {
-          emit("time-entry-selected", item.value);
-        },
-        // Because the autocomplete plugion always places elements
-        // at the bottom we use Popper.js to update its position.
-        customize: (input: HTMLInputElement, _, container: HTMLDivElement) => {
-          // ToDo: using Popper.js in this way (destroying, creating) is not poptimal
-          if (popper !== undefined) {
-            popper.destroy();
+        render(item: AutocompleteItem): HTMLDivElement | undefined {
+          if (item.type === "time-entry") {
+            return renderAutocompleteTimeEntryEntry(item);
           }
 
-          popper = createPopper(input, container, {
-            placement: "auto",
-            modifiers: [
-              {
-                name: "offset",
-                options: {
-                  offset: [0, 10],
-                },
-              },
-            ],
-          });
+          if (item.type === "customer") {
+            return renderAutocompleteCustomerEntry(item);
+          }
+
+          if (item.type === "project") {
+            return renderAutocompleteProjectEntry(
+              item as ProjectAutocompleteItem
+            );
+          }
+        },
+        onSelect(item: AutocompleteItem) {
+          if (item.type === "time-entry") {
+            emit("time-entry-selected", item.value);
+          }
+
+          if (item.type === "customer") {
+            emit("customer-selected", item.value);
+            input.value.value = input.value.value.slice(0, spliceIndex).trim();
+            modelProxy.value.description = input.value.value;
+            spliceIndex = undefined;
+          }
+
+          if (item.type === "project") {
+            emit("project-selected", item.value);
+            input.value.value = input.value.value.slice(0, spliceIndex).trim();
+            modelProxy.value.description = input.value.value;
+            spliceIndex = undefined;
+          }
+        },
+        customize: (
+          input: HTMLInputElement,
+          inputRect,
+          container: HTMLDivElement,
+          maxHeight: number
+        ) => {
+          container.style.width = `${
+            input.parentElement?.getBoundingClientRect().width
+          }px`;
+
+          container.style.visibility = "visible";
+          container.style.zIndex = "1000";
+
+          if (maxHeight < 100) {
+            container.style.top = "";
+            container.style.bottom =
+              window.innerHeight - inputRect.bottom + input.offsetHeight + "px";
+            container.style.maxHeight = "200px";
+          }
         },
       });
     });