+++ src/directives/autofocus.ts
@@ -2,7 +2,6 @@ import { Directive } from "vue";
 
 const focus: Directive = {
   mounted(el: HTMLElement) {
-    console.log(el);
     el.focus();
   },
 };